fix: consolidate named factory list/delete for pkg-file-count
Merge ListNamedFactories and DeleteNamedFactory into runtime_config.go
so pkg/config stays within the 15-file package limit required by CI.
