+++ docs/internal/processes/development-guide-relevant-files.md
@@ -525,10 +525,10 @@ This inventory records the checked-in files and directories that the maintainer 
 - Keep `GET /work` explicit sort fields narrow and contract-owned: add public sort options to `api/openapi-main.yaml`, regenerate API artifacts, validate unsupported `sortBy` values in `pkg/api/handlers.go`, and route CLI flags such as `--sort-by` to the exact API query value.
 - Keep `GET /work` pagination cursor-based with `maxResults`, `nextToken`, and an always-present `paginationContext`; omit only `paginationContext.nextToken` when the current page is final or trailing-empty so CLI JSON automation can distinguish page metadata from continuation. When the public list has semantic ordering, resume by locating the exact token-ID cursor in that ordered list rather than comparing token IDs lexically.
 - Keep `pkg/cli/work/list.go` as the owner of `work list` presentation: default output is a human-readable tabular view with work identity and state columns, while `--json` should continue to encode the generated `ListWorkResponse` directly for automation.
-- Keep persisted named-factory discovery in `pkg/config/named_factory_list.go` via `ListNamedFactories(rootDir)`; scan factory-root subdirectories with `requireFactoryConfig`, mark `.current-factory` with `ReadCurrentFactoryPointer`, and let `pkg/cli/factory/list.go` own human-readable table output (`NAME`, `FACTORY DIRECTORY`, `CURRENT`) plus JSON array encoding for scripting.
+- Keep persisted named-factory discovery in `pkg/config/runtime_config.go` via `ListNamedFactories(rootDir)`; scan factory-root subdirectories with `requireFactoryConfig`, mark `.current-factory` with `ReadCurrentFactoryPointer`, and let `pkg/cli/factory/list.go` own human-readable table output (`NAME`, `FACTORY DIRECTORY`, `CURRENT`) plus JSON array encoding for scripting.
 - File-based named-factory create flows should default `--dir` to `defaultcmd.FactoryDir`, read the `--from` factory.json payload in `pkg/cli/factory/save.go`, delegate persistence to `config.PersistNamedFactory`, optionally update `.current-factory` with `WriteCurrentFactoryPointer` when `--set-current` is set, and keep human-readable success output separate from `--json` confirmation objects.
 - File-based named-factory replace flows should mirror save flags in `pkg/cli/factory/update.go`, delegate to `config.ReplaceNamedFactory`, map `os.ErrNotExist` to `factory not found` and `ErrInvalidNamedFactory` to `invalid factory config`, and leave `.current-factory` untouched so replacing the current named factory keeps the same pointer.
-- File-based named-factory delete flows should default `--dir` to `defaultcmd.FactoryDir`, delegate removal to `config.DeleteNamedFactory` in `pkg/config/named_factory_delete.go`, map `os.ErrNotExist` to `factory not found` and `ErrNamedFactoryIsCurrent` to `cannot delete current factory`, and never remove `.current-factory` when deleting a non-current name.
+- File-based named-factory delete flows should default `--dir` to `defaultcmd.FactoryDir`, delegate removal to `config.DeleteNamedFactory` in `pkg/config/runtime_config.go`, map `os.ErrNotExist` to `factory not found` and `ErrNamedFactoryIsCurrent` to `cannot delete current factory`, and never remove `.current-factory` when deleting a non-current name.
 - Live `you factory save` (no name argument) should default `--port` to `defaultcmd.FactoryPort` and route GET+PUT through `sessionpath.CurrentFactoryPath` in `pkg/cli/factory/save_current.go`, reusing `queryCurrent` for the read and echoing server `version` metadata in the PUT body; surface API `ErrorResponse` messages on validation or conflict failures and keep `--json` stdout limited to the saved `Factory` payload.
 - Keep `you factory` parent help, README factory examples, and `pkg/cli/root_factory_test.go` aligned when adding factory subcommands: parent `Long` should enumerate query/list/save/update/delete, `Example` should show one representative command per surface, and registration tests should `Find` each subcommand plus assert `factory --help` mentions the installed `you` examples.
 - Keep the reserved default-runtime current-factory identifier documented in `api/components/schemas/data-models/FactoryName.yaml` and `/factory/~current`; regenerate bundled artifacts after authored OpenAPI edits instead of hand-patching `api/openapi.yaml` or generated clients.
